+++ src/npg_porch/db/data_access.py
@@ -28,8 +28,10 @@ from sqlalchemy.orm import contains_eager, joinedload
 from sqlalchemy.orm.exc import NoResultFound
 from sqlalchemy.sql.functions import count, max as samax
 
+from npg_porch.db.exceptions import IncompatibleArgumentsException
 from npg_porch.db.models import Event
 from npg_porch.db.models import Pipeline as DbPipeline
+from npg_porch.db.models import Version as DbVersion
 from npg_porch.db.models import Task as DbTask
 from npg_porch.db.models import Token as DbToken
 from npg_porch.models import Pipeline, Task, TaskStateEnum, TaskExpanded
@@ -62,10 +64,28 @@ class AsyncDbAccessor:
     async def _get_pipeline_db_objects(
         self,
         name: str | None = None,
-        version: str | None = None,
         uri: str | None = None,
     ) -> list[Pipeline]:
         query = select(DbPipeline)
+        if name:
+            query = query.filter_by(name=name)
+        if uri:
+            query = query.filter_by(repository_uri=uri)
+
+        pipeline_result = await self.session.execute(query)
+        return pipeline_result.scalars().all()
+
+    async def _get_pipeline_version_db_objects(
+        self,
+        name: str | None = None,
+        version: str | None = None,
+        uri: str | None = None,
+    ):
+        query = (
+            select(DbVersion)
+            .join(DbVersion.pipeline)
+            .options(joinedload(DbVersion.pipeline))
+        )
         if name:
             query = query.filter_by(name=name)
         if version:
@@ -73,14 +93,33 @@ class AsyncDbAccessor:
         if uri:
             query = query.filter_by(repository_uri=uri)
 
-        pipeline_result = await self.session.execute(query)
-        return pipeline_result.scalars().all()
+        version_result = await self.session.execute(query)
+        return version_result.scalars().all()
 
-    async def get_all_pipelines(
-        self, uri: str | None = None, version: str | None = None
-    ) -> list[Pipeline]:
-        pipelines = await self._get_pipeline_db_objects(uri=uri, version=version)
+    async def _get_version_db_object(
+        self,
+        name: str,
+        version: str,
+    ):
+        version_result = await self.session.execute(
+            (
+                select(DbVersion)
+                .join(DbVersion.pipeline)
+                .options(joinedload(DbVersion.pipeline))
+            )
+            .filter(DbPipeline.name == name)
+            .filter(DbVersion.version == version)
+        )
+
+        return version_result.scalar_one()
+
+    async def get_all_pipelines(self, uri: str | None = None) -> list[Pipeline]:
+        pipelines = await self._get_pipeline_db_objects(uri=uri)
         return [pipe.convert_to_model() for pipe in pipelines]
+
+    async def get_pipeline_versions(self, pipeline_name: str) -> list[str]:
+        versions = await self._get_pipeline_version_db_objects(name=pipeline_name)
+        return [version.version for version in versions]
 
     async def get_recent_pipelines(self):
         query = (
@@ -95,15 +134,32 @@ class AsyncDbAccessor:
 
     async def create_pipeline(self, pipeline: Pipeline) -> Pipeline:
         session = self.session
-        assert isinstance(pipeline, Pipeline)
+        if not isinstance(pipeline, Pipeline):
+            raise TypeError
 
-        pipe = DbPipeline(
-            name=pipeline.name, version=pipeline.version, repository_uri=pipeline.uri
+        pipe = DbPipeline(name=pipeline.name, repository_uri=pipeline.uri)
+        ver = DbVersion(version=pipeline.version, pipeline=pipe)
+        session.add(ver)
+        await session.commit()
+        return ver.convert_to_api_pipeline()
+
+    async def create_version(self, pipeline: Pipeline) -> Pipeline:
+        session = self.session
+        if not isinstance(pipeline, Pipeline):
+            raise TypeError
+
+        try:
+            pipe = await self._get_pipeline_db_object(pipeline.name)
+        except NoResultFound:
+            raise NoResultFound("Pipeline not found")
+        ver = DbVersion(
+            pipeline=pipe,
+            version=pipeline.version,
         )
 
-        session.add(pipe)
+        session.add(ver)
         await session.commit()
-        return pipe.convert_to_model()
+        return ver.convert_to_api_pipeline()
 
     async def create_pipeline_token(self, name: str, desc: str) -> Token:
         session = self.session
@@ -126,10 +182,12 @@ class AsyncDbAccessor:
         """
         self.logger.debug("CREATE TASK: " + str(task))
         session = self.session
-        db_pipeline = await self._get_pipeline_db_object(task.pipeline.name)
+        db_version = await self._get_version_db_object(
+            version=task.pipeline.version, name=task.pipeline.name
+        )
 
         task.status = TaskStateEnum.PENDING
-        t = self.convert_task_to_db(task, db_pipeline)
+        t = self.convert_task_to_db(task, db_version)
         created = True
         try:
             nested = await session.begin_nested()
@@ -142,11 +200,13 @@ class AsyncDbAccessor:
             # Task already exists, query the database to get the up-to-date
             # representation of the task.
             t = await self.get_db_task(
-                pipeline_name=task.pipeline.name, job_descriptor=t.job_descriptor
+                pipeline_name=task.pipeline.name,
+                version=task.pipeline.version,
+                job_descriptor=t.job_descriptor,
             )
             created = False
 
-        return (t.convert_to_model(), created)
+        return t.convert_to_model(), created
 
     async def claim_tasks(
         self, token_id: int, pipeline: Pipeline, claim_limit: int | None = 1
@@ -154,20 +214,20 @@ class AsyncDbAccessor:
         session = self.session
 
         try:
-            pipeline_result = await session.execute(
-                select(DbPipeline).filter_by(name=pipeline.name)
+            version = await self._get_version_db_object(
+                name=pipeline.name, version=pipeline.version
             )
-            pipeline = pipeline_result.scalar_one()
         except NoResultFound:
             raise NoResultFound("Pipeline not found")
 
         potential_tasks = await session.execute(
             select(DbTask)
-            .join(DbTask.pipeline)
-            .where(DbPipeline.name == pipeline.name)
+            .join(DbTask.pipeline_version)
+            .join(DbVersion.pipeline)
+            .where(DbVersion.version_id == version.version_id)
             .where(DbTask.state == TaskStateEnum.PENDING)
             .order_by(DbTask.created)
-            .options(contains_eager(DbTask.pipeline))
+            .options(contains_eager(DbTask.pipeline_version, DbVersion.pipeline))
             .with_for_update()
             .limit(claim_limit)
             .execution_options(populate_existing=True)
@@ -198,14 +258,16 @@ class AsyncDbAccessor:
         session = self.session
         # Get the matching task from the DB
         try:
-            db_pipe = await self._get_pipeline_db_object(task.pipeline.name)
+            db_ver = await self._get_version_db_object(
+                name=task.pipeline.name, version=task.pipeline.version
+            )
         except NoResultFound:
             raise NoResultFound("Pipeline not found")
 
         try:
             task_result = await session.execute(
                 select(DbTask)
-                .filter_by(pipeline=db_pipe)
+                .filter_by(pipeline_version=db_ver)
                 .filter_by(job_descriptor=task.generate_task_id())
             )
             og_task = task_result.scalar_one()  # Doesn't raise exception if no rows?!
@@ -227,22 +289,35 @@ class AsyncDbAccessor:
         return og_task.convert_to_model()
 
     async def get_tasks(
-        self, pipeline_name: str | None = None, task_status: TaskStateEnum | None = None
+        self,
+        pipeline_name: str | None = None,
+        task_status: TaskStateEnum | None = None,
+        version: str | None = None,
     ) -> list[Task]:
         """
         Gets all the tasks.
 
-        Can filter tasks by pipeline name and task status in order to be more useful.
+        Can filter tasks by pipeline name, task status and version in order to be more useful.
         """
+        if version and not pipeline_name:
+            raise IncompatibleArgumentsException(
+                "A version without a pipeline name is not meaningful when getting tasks"
+            )
         query = (
-            select(DbTask).join(DbTask.pipeline).options(joinedload(DbTask.pipeline))
+            select(DbTask)
+            .join(DbTask.pipeline_version)
+            .join(DbVersion.pipeline)
+            .options(contains_eager(DbTask.pipeline_version, DbVersion.pipeline))
         )
 
         if pipeline_name:
             query = query.where(DbPipeline.name == pipeline_name)
 
         if task_status:
-            query = query.filter(DbTask.state == task_status)
+            query = query.where(DbTask.state == task_status)
+
+        if version:
+            query = query.where(DbVersion.version == version)
 
         task_result = await self.session.execute(query)
         tasks = task_result.scalars().all()
@@ -253,6 +328,7 @@ class AsyncDbAccessor:
         pipeline_name: str = None,
         status: list[TaskStateEnum] = None,
         since: datetime = None,
+        version: str = None,
     ) -> list[TaskExpanded]:
         """
         Gets information about tasks including their creation date, ordered
@@ -260,6 +336,10 @@ class AsyncDbAccessor:
 
         Can be filtered by pipeline name and status.
         """
+        if version and not pipeline_name:
+            raise IncompatibleArgumentsException(
+                "A version without a pipeline name is not meaningful when getting tasks"
+            )
         latest_event = (
             select(samax(Event.time).label("status_date"), Event.task_id)
             .select_from(Event)
@@ -270,8 +350,9 @@ class AsyncDbAccessor:
             select(DbTask, latest_event.c.status_date)
             .select_from(DbTask)
             .join(latest_event, DbTask.task_id == latest_event.c.task_id)
-            .join(DbTask.pipeline)
-            .options(contains_eager(DbTask.pipeline))
+            .join(DbTask.pipeline_version)
+            .join(DbVersion.pipeline)
+            .options(contains_eager(DbTask.pipeline_version, DbVersion.pipeline))
             .order_by(latest_event.c.status_date.desc())
         )
 
@@ -281,6 +362,8 @@ class AsyncDbAccessor:
             query = query.where(or_(DbTask.state == state for state in status))
         if since:
             query = query.where(latest_event.c.status_date >= since)
+        if version:
+            query = query.where(DbVersion.version == version)
 
         self.logger.debug(query.compile())
         task_result = await self.session.execute(query)
@@ -303,11 +386,13 @@ class AsyncDbAccessor:
         lengths = {}
         not_done = {}
         for task in tasks:
+            if task.status in [TaskStateEnum.CANCELLED, TaskStateEnum.FAILED]:
+                continue
             if task.status == TaskStateEnum.DONE:
                 lengths.setdefault(task.pipeline.name, []).append(
                     (task.updated - task.created).total_seconds()
                 )
-            elif task.status != TaskStateEnum.CANCELLED:
+            else:
                 not_done.setdefault(task.pipeline.name, []).append(task)
         long_running = []
         for pipeline, pipeline_tasks in not_done.items():
@@ -323,13 +408,16 @@ class AsyncDbAccessor:
         self,
         pipeline_name: str,
         job_descriptor: str,
+        version: str,
     ) -> DbTask:
         """Get the task."""
         query = (
             select(DbTask)
-            .join(DbTask.pipeline)
-            .options(joinedload(DbTask.pipeline))
+            .join(DbTask.pipeline_version)
+            .join(DbVersion.pipeline)
+            .options(joinedload(DbTask.pipeline_version, DbVersion.pipeline))
             .where(DbPipeline.name == pipeline_name)
+            .where(DbVersion.version == version)
             .where(DbTask.job_descriptor == job_descriptor)
         )
         task_result = await self.session.execute(query)
@@ -337,11 +425,12 @@ class AsyncDbAccessor:
         return task_result.scalars().one()
 
     @staticmethod
-    def convert_task_to_db(task: Task, pipeline: DbPipeline) -> DbTask:
-        assert task.status in TaskStateEnum
+    def convert_task_to_db(task: Task, version: DbVersion) -> DbTask:
+        if task.status not in TaskStateEnum:
+            raise ValueError
 
         return DbTask(
-            pipeline=pipeline,
+            pipeline_version=version,
             job_descriptor=task.generate_task_id(),
             definition=task.task_input,
             state=task.status,